test(cketh): fold the sweep e2e into the deposit_from_cex target
The twenty-deposit sweep test had its own file and Bazel target, on the
theory that sharing a sequential target would starve the other live
tests of CPU. It doesn't: the merged target runs all six tests in 81
seconds. One file and one target now carry the whole deposit-from-CEX
story — batcher semantics, balance scan, sweeper funding, and the sweep
itself — with the target sized large for the longer run.
