+++ rs/ethereum/cketh/minter/BUILD.bazel
@@ -209,7 +209,6 @@ rust_ic_test_suite(
             # Run against a local foundry anvil node, not the IC test suite.
             "tests/deposit_from_cex_demo.rs",
             "tests/deposit_from_cex.rs",
-            "tests/deposit_from_cex_sweep.rs",
             "tests/sweeper_funding.rs",
         ],
     ),
@@ -294,59 +293,11 @@ rust_test(
 
 rust_test(
     name = "deposit_from_cex",
-    # The end-to-end balance-scan test runs a live PocketIC and waits (wall-clock) for the minter's
-    # periodic scan, so it needs a larger timeout than the pure-anvil tests.
-    size = "medium",
-    srcs = ["tests/deposit_from_cex.rs"],
-    data = [
-        "//:anvil",
-        "//:solc",
-        # Reuses the mock ERC-20 source compiled at test time by the vendored solc.
-        "DepositHelperWithSubaccount.sol",
-        "tests/deposit_from_cex_demo/MockUSDT.sol",
-        # End-to-end balance scan and sweeper funding on a live PocketIC + local anvil node.
-        ":cketh_minter_debug.wasm.gz",
-        "//rs/ethereum/ledger-suite-orchestrator:ledger_suite_orchestrator_canister.wasm.gz",
-        "//rs/ledger_suite/icrc1/archive:archive_canister_u256.wasm.gz",
-        "//rs/ledger_suite/icrc1/index-ng:index_ng_canister_u256.wasm.gz",
-        "//rs/ledger_suite/icrc1/ledger:ledger_canister_u256.wasm.gz",
-        "//rs/pocket_ic_server:pocket-ic-server",
-        "@evm_rpc.wasm.gz//file",
-    ],
-    env = {
-        "ANVIL_BIN": "$(rootpath //:anvil)",
-        "CKDEPOSIT_SOL": "$(rootpath DepositHelperWithSubaccount.sol)",
-        # One test at a time: two live PocketIC instances and their anvil nodes running
-        # concurrently starve each other of CPU, and an ingress message that loses that race is
-        # stamped with an expiry the auto-progressing clock then moves past, never to be answered.
-        "RUST_TEST_THREADS": "1",
-        "MOCKUSDT_SOL": "$(rootpath tests/deposit_from_cex_demo/MockUSDT.sol)",
-        "SOLC_BIN": "$(rootpath //:solc)",
-        "CARGO_MANIFEST_DIR": "rs/ethereum/cketh/minter",
-        "CKETH_MINTER_WASM_PATH": "$(rootpath :cketh_minter_debug.wasm.gz)",
-        "EVM_RPC_CANISTER_WASM_PATH": "$(rootpath @evm_rpc.wasm.gz//file)",
-        "INDEX_CANISTER_WASM_PATH": "$(rootpath //rs/ledger_suite/icrc1/index-ng:index_ng_canister_u256.wasm.gz)",
-        "LEDGER_ARCHIVE_NODE_CANISTER_WASM_PATH": "$(rootpath //rs/ledger_suite/icrc1/archive:archive_canister_u256.wasm.gz)",
-        "LEDGER_CANISTER_WASM_PATH": "$(rootpath //rs/ledger_suite/icrc1/ledger:ledger_canister_u256.wasm.gz)",
-        "LEDGER_SUITE_ORCHESTRATOR_WASM_PATH": "$(rootpath //rs/ethereum/ledger-suite-orchestrator:ledger_suite_orchestrator_canister.wasm.gz)",
-        "POCKET_IC_BIN": "$(rootpath //rs/pocket_ic_server:pocket-ic-server)",
-    },
-    deps = [
-        # Keep sorted.
-        ":minter",
-        "//packages/ic-ethereum-types",
-        "//rs/ethereum/cketh/test_utils",
-        "@crate_index//:assert_matches",
-    ],
-)
-
-rust_test(
-    name = "deposit_from_cex_sweep",
-    # Its own target rather than another case in `deposit_from_cex`: this drives twenty deposits
-    # through two sweeps on one live PocketIC, and sharing a (sequential) target with the other
-    # end-to-end tests starves them of CPU until their ingress messages expire unanswered.
+    # The end-to-end tests each run a live PocketIC and wait (wall-clock) for the minter's periodic
+    # tasks — and the sweep test drives twenty deposits through two sweeps — so the target needs a
+    # far larger timeout than the pure-anvil tests.
     size = "large",
-    srcs = ["tests/deposit_from_cex_sweep.rs"],
+    srcs = ["tests/deposit_from_cex.rs"],
     data = [
         "//:anvil",
         "//:solc",
